✨ improve(patch): bud-stylelint api
